third_party/WebKit/LayoutTests/accessibility/aria-hidden-updates-alldescendants-expected.txt

Steps
test
Step 1: Do something
Step 2: Do another thing
Step 3: Do one last thing

This tests that if aria-hidden changes on an element, all it's existing children will update their children caches

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS main.childrenCount is 1
PASS main.childrenCount is 2
PASS main.childAtIndex(1).childrenCount is 1
PASS successfullyParsed is true

TEST COMPLETE
